+++ server/src/db/create_azuread_table.sql
@@ -3,6 +3,7 @@ DROP TABLE IF EXISTS `azuread`;
 CREATE TABLE `azuread` (
   `client_id` text DEFAULT NULL,
   `secret_id` text DEFAULT NULL,
-  `enable` tinyint(4) DEFAULT NULL
+  `enable` tinyint(4) DEFAULT NULL,
+  `groupfilter` varchar(250) DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
-INSERT INTO AnsibleForms.azuread(client_id,secret_id,enable) VALUES('','',0);
+INSERT INTO AnsibleForms.azuread(client_id,secret_id,enable,groupfilter) VALUES('','',0,'');